pyproject.toml
[project]
name = "colorfinder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
]
